feat: add `SQLData` class
