+++ common/src/main/java/com/hibiscusmc/hmccosmetics/database/types/MySQLData.java
@@ -2,19 +2,14 @@ package com.hibiscusmc.hmccosmetics.database.types;
 
 import com.hibiscusmc.hmccosmetics.HMCCosmeticsPlugin;
 import com.hibiscusmc.hmccosmetics.config.DatabaseSettings;
-import com.hibiscusmc.hmccosmetics.cosmetic.Cosmetic;
-import com.hibiscusmc.hmccosmetics.cosmetic.CosmeticSlot;
-import com.hibiscusmc.hmccosmetics.user.CosmeticUser;
 import org.bukkit.Bukkit;
-import org.bukkit.Color;
 import org.jetbrains.annotations.NotNull;
 
 import java.sql.*;
-import java.util.Map;
 import java.util.Properties;
 import java.util.UUID;
 
-public class MySQLData extends Data {
+public class MySQLData extends SQLData {
 
     // Connection Information
     private String host;
@@ -53,51 +48,6 @@ public class MySQLData extends Data {
             throw new RuntimeException(e);
         }
     }
-    @Override
-    public void save(CosmeticUser user) {
-        Runnable run = () -> {
-            try (PreparedStatement preparedSt = preparedStatement("REPLACE INTO COSMETICDATABASE(UUID,COSMETICS) VALUES(?,?);")) {
-                preparedSt.setString(1, user.getUniqueId().toString());
-                preparedSt.setString(2, serializeData(user));
-                preparedSt.executeUpdate();
-            } catch (SQLException e) {
-                throw new RuntimeException(e);
-            }
-        };
-        if (!HMCCosmeticsPlugin.isDisable()) {
-            Bukkit.getScheduler().runTaskAsynchronously(HMCCosmeticsPlugin.getInstance(), run);
-        } else {
-            run.run();
-        }
-    }
-
-    @Override
-    public CosmeticUser get(UUID uniqueId) {
-        CosmeticUser user = new CosmeticUser(uniqueId);
-
-        Bukkit.getScheduler().runTaskAsynchronously(HMCCosmeticsPlugin.getInstance(), () -> {
-            try {
-                PreparedStatement preparedStatement = preparedStatement("SELECT * FROM COSMETICDATABASE WHERE UUID = ?;");
-                preparedStatement.setString(1, uniqueId.toString());
-                ResultSet rs = preparedStatement.executeQuery();
-                if (rs.next()) {
-                    String rawData = rs.getString("COSMETICS");
-                    Map<CosmeticSlot, Map<Cosmetic, Color>> cosmetics = deserializeData(user, rawData);
-                    for (Map<Cosmetic, Color> cosmeticColors : cosmetics.values()) {
-                        for (Cosmetic cosmetic : cosmeticColors.keySet()) {
-                            Bukkit.getScheduler().runTask(HMCCosmeticsPlugin.getInstance(), () -> {
-                                // This can not be async.
-                                user.addPlayerCosmetic(cosmetic, cosmeticColors.get(cosmetic));
-                            });
-                        }
-                    }
-                }
-            } catch (SQLException e) {
-                e.printStackTrace();
-            }
-        });
-        return user;
-    }
 
     @Override
     public void clear(UUID uniqueId) {
@@ -110,7 +60,8 @@ public class MySQLData extends Data {
                 e.printStackTrace();
             }
         });
-        // TODO
+
+        // TODO (Something)
     }
 
     private void openConnection() throws SQLException {
@@ -162,6 +113,7 @@ public class MySQLData extends Data {
         }
     }
 
+    @Override
     public PreparedStatement preparedStatement(String query) {
         PreparedStatement ps = null;
