+++ common/src/main/java/com/hibiscusmc/hmccosmetics/database/types/SQLiteData.java
@@ -1,21 +1,16 @@
 package com.hibiscusmc.hmccosmetics.database.types;
 
 import com.hibiscusmc.hmccosmetics.HMCCosmeticsPlugin;
-import com.hibiscusmc.hmccosmetics.cosmetic.Cosmetic;
-import com.hibiscusmc.hmccosmetics.cosmetic.CosmeticSlot;
-import com.hibiscusmc.hmccosmetics.user.CosmeticUser;
 import com.hibiscusmc.hmccosmetics.util.MessagesUtil;
 import org.bukkit.Bukkit;
-import org.bukkit.Color;
 
 import java.io.File;
 import java.io.IOException;
 import java.sql.*;
-import java.util.Map;
 import java.util.UUID;
 import java.util.logging.Level;
 
-public class SQLiteData extends Data {
+public class SQLiteData extends SQLData {
 
     private Connection connection;
 
@@ -39,58 +34,9 @@ public class SQLiteData extends Data {
                     "(UUID varchar(36) PRIMARY KEY, " +
                     "COSMETICS MEDIUMTEXT " +
                     ");").execute();
-
-
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
-    }
-
-    @Override
-    public void save(CosmeticUser user) {
-        Runnable run = () -> {
-            try {
-                PreparedStatement preparedSt = preparedStatement("REPLACE INTO COSMETICDATABASE(UUID,COSMETICS) VALUES(?,?);");
-                preparedSt.setString(1, user.getUniqueId().toString());
-                preparedSt.setString(2, serializeData(user));
-                preparedSt.executeUpdate();
-            } catch (SQLException e) {
-                throw new RuntimeException(e);
-            }
-        };
-        if (!HMCCosmeticsPlugin.isDisable()) {
-            Bukkit.getScheduler().runTaskAsynchronously(HMCCosmeticsPlugin.getInstance(), run);
-        } else {
-            run.run();
-        }
-    }
-
-    @Override
-    public CosmeticUser get(UUID uniqueId) {
-        CosmeticUser user = new CosmeticUser(uniqueId);
-
-        Bukkit.getScheduler().runTaskAsynchronously(HMCCosmeticsPlugin.getInstance(), () -> {
-            try {
-                PreparedStatement preparedStatement = preparedStatement("SELECT * FROM COSMETICDATABASE WHERE UUID = ?;");
-                preparedStatement.setString(1, uniqueId.toString());
-                ResultSet rs = preparedStatement.executeQuery();
-                if (rs.next()) {
-                    String rawData = rs.getString("COSMETICS");
-                    Map<CosmeticSlot, Map<Cosmetic, Color>> cosmetics = deserializeData(user, rawData);
-                    for (Map<Cosmetic, Color> cosmeticColors : cosmetics.values()) {
-                        for (Cosmetic cosmetic : cosmeticColors.keySet()) {
-                            Bukkit.getScheduler().runTask(HMCCosmeticsPlugin.getInstance(), () -> {
-                                // This can not be async.
-                                user.addPlayerCosmetic(cosmetic, cosmeticColors.get(cosmetic));
-                            });
-                        }
-                    }
-                }
-            } catch (SQLException e) {
-                e.printStackTrace();
-            }
-        });
-        return user;
     }
 
     @Override
@@ -104,7 +50,6 @@ public class SQLiteData extends Data {
             }
         });
     }
-
 
     private void openConnection() throws SQLException {
         // Bukkit.getScheduler().runTaskAsynchronously(HMCCosmeticsPlugin.getInstance(), () -> {
@@ -126,16 +71,19 @@ public class SQLiteData extends Data {
         }
     }
 
+    @Override
     public PreparedStatement preparedStatement(String query) {
         PreparedStatement ps = null;
         if (!isConnectionOpen()) {
             HMCCosmeticsPlugin.getInstance().getLogger().info("Connection is not open");
         }
+
         try {
             ps = connection.prepareStatement(query);
         } catch (SQLException e) {
             e.printStackTrace();
         }
+
         return ps;
     }
